XML-to-RDF/XML scripts
Here is a collection of three scripts. One to count up the number of output files which had DOI/PubMed ID errors. One to read and write the data files to new folders (I see a typo in the file name, should be data not date). And the main script that goes through the directory, reads the old files and makes new SBOL-compliant RDF/XML files (i.e., createSBOL_XML.py).